Renamed project package. Added firebase dependencies.
