+++ app/google-services.json
@@ -1,0 +1,29 @@
+{
+  "project_info": {
+    "project_number": "641749093332",
+    "project_id": "pokerapp-8f562",
+    "storage_bucket": "pokerapp-8f562.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:641749093332:android:e4179babd6baf318a9e441",
+        "android_client_info": {
+          "package_name": "com.example.pokerapp"
+        }
+      },
+      "oauth_client": [],
+      "api_key": [
+        {
+          "current_key": "AIzaSyCfJ9YCIhLwaa5wrFrjuUQAkhtPsTZ4jFo"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": []
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
