+++ app/src/main/java/com/example/pokerapp/PokerApp.kt
@@ -1,4 +1,4 @@
-package com.example.projectapp
+package com.example.pokerapp
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
