+++ app/src/main/java/com/example/pokerapp/model/RealtimeMessagingClient.kt
@@ -1,4 +1,4 @@
-package com.example.projectapp.model
+package com.example.pokerapp.model
 
 import kotlinx.coroutines.flow.Flow
