+++ app/src/main/java/com/example/pokerapp/model/User.kt
@@ -1,4 +1,4 @@
-package com.example.projectapp.model
+package com.example.pokerapp.model
 
 class User(val username: String, var chipAmount: Int = 1000) {
     var avatar: String? = null
