+++ app/src/main/java/com/example/pokerapp/ui/GameViewModel.kt
@@ -1,6 +1,5 @@
-package com.example.projectapp.ui
+package com.example.pokerapp.ui
 
-import android.util.Log
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableIntStateOf
@@ -8,18 +7,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.projectapp.model.GameState
-import com.example.projectapp.model.PlayerAction
-import com.example.projectapp.model.RealtimeMessagingClient
+import com.example.pokerapp.model.GameState
+import com.example.pokerapp.model.PlayerAction
+import com.example.pokerapp.model.RealtimeMessagingClient
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.stateIn
