+++ app/src/main/java/com/example/pokerapp/ui/TableInformation.kt
@@ -1,4 +1,4 @@
-package com.example.projectapp.ui
+package com.example.pokerapp.ui
 
 import android.content.Context
 import androidx.compose.foundation.Image
@@ -21,8 +21,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.projectapp.R
-import com.example.projectapp.model.GameState
+import com.example.pokerapp.R
+import com.example.pokerapp.model.GameState
 
 @Composable
 fun TableBackground(){
