+++ build.gradle.kts
@@ -14,4 +14,5 @@ plugins {
     alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.jetbrainsKotlinAndroid) apply false
     id("com.google.dagger.hilt.android") version "2.46" apply false
+    id("com.google.gms.google-services") version "4.4.1" apply false
 }
